bug(properties): move the validates to the middleware[finishes #167181853]
