+++ src/controllers/PropertyController.js
@@ -1,6 +1,6 @@
 
-import propertyId from '../../database/Property';
-import reportId from '../../database/report';
+import propertyId from '../database/Property';
+import reportId from '../database/report';
 import date from '../utils/helpers/dates';
 import Uid from '../utils/helpers/Ids';
 import PropertyModel from '../Models/PropertyModel';
@@ -13,8 +13,9 @@ export default class Property {
     static async Post(req, res) {
         try {
             const {
-                price, state, city, address, type,
+                state, city, address, type,
             } = req.body;
+            const { price } = res.locals;
             const id = Uid(propertyId);
             const imageUrl = await upload(req, id);
             const owner = res.locals.user;
@@ -59,7 +60,6 @@ export default class Property {
             const id = parseInt(req.params.property_id, 10);
             const property = new PropertyModel(id);
             if (await property.findOne()) return Res.handleSuccess(200, 'got property successfully', property.result, res);
-            return Res.handleError(404, 'Property with such id does not exists', res);
         } catch (err) {
             return Res.handleError(500, err.toString(), res);
         }
@@ -67,9 +67,7 @@ export default class Property {
 
     static async markSold(req, res) {
         try {
-            const id = parseInt(req.params.property_id, 10);
-            const owner = res.locals.user;
-            if (!await PropertyModel.checkUser(id, owner)) return Res.handleError(406, 'None of the ads with such id belongs to you', res);
+            const { id } = res.locals;
             const sold = { status: 'sold', id };
             const property = new PropertyModel(sold);
             await property.update();
@@ -81,9 +79,7 @@ export default class Property {
 
     static async delProperty(req, res) {
         try {
-            const id = parseInt(req.params.property_id, 10);
-            const owner = res.locals.user;
-            if (!await PropertyModel.checkUser(id, owner)) return Res.handleError(406, 'None of the ads with such id belongs to you', res);
+            const { id } = res.locals;
             const property = new PropertyModel(id);
             await property.del();
             // nextline handleError works as handleSuccess
@@ -97,7 +93,7 @@ export default class Property {
         try {
             const { type } = req.query;
             const property = new PropertyModel(type);
-            if (!await property.getType()) return Res.handleError(404, 'adverts with this type does not exists', res);
+            await property.getType();
             return Res.handleSuccess(200, 'got specific type Successful', property.result, res);
         } catch (err) {
             return Res.handleError(500, err.toString(), res);
@@ -116,7 +112,7 @@ export default class Property {
             const property = new PropertyModel({
                 reason, description, id, createdOn, propertyId,
             });
-            if (!await property.report()) return Res.handleError(404, 'Property with such id does not exists', res);
+            await property.report();
             return Res.handleSuccess(201, 'successfully created a report', property.result, res);
         } catch (err) {
             return Res.handleError(500, err.toString(), res);
