+++ src/database/Property.js
@@ -1,0 +1,3 @@
+const Property = [];
+
+export default Property;
